% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getModelResults}
\alias{getModelResults}
\title{Extract the coefficients and standard errors from a data frame of estimated models}
\usage{
getModelResults(models)
}
\arguments{
\item{models}{A data frame containing estimated models for a series of
different sample sizes.}
}
\value{
A data frame of the estimated coefficients and standard errors for
each parameter in each model.
}
\description{
Returns a data frame of the estimated coefficients and standard errors for
each parameter in each model in a data frame of estimated models.
}
\examples{
\dontrun{
library(conjointTools)

# Define the attributes and levels
levels <- list(
  price     = seq(1, 4, 0.5), # $ per pound
  type      = c('Fuji', 'Gala', 'Honeycrisp', 'Pink Lady', 'Red Delicious'),
  freshness = c('Excellent', 'Average', 'Poor')
)

# Make a full-factorial design of experiment and recode the levels
doe <- makeDoe(levels)
doe <- recodeDesign(doe, levels)

# Make the survey
survey <- makeSurvey(
    doe       = doe,  # Design of experiment
    nResp     = 2000, # Total number of respondents (upper bound)
    nAltsPerQ = 3,    # Number of alternatives per question
    nQPerResp = 6     # Number of questions per respondent
)

# Simulate random choices for the survey
data <- simulateChoices(
    survey    = survey,
    altIDName = "altID",
    obsIDName = "obsID"
)

# Estimate models with different sample sizes
models <- estimateModels(
    nbreaks     = 10,
    data        = data,
    parNames    = c("price", "type", "freshness"),
    choiceName  = "choice",
    obsIDName   = "obsID"
)

# Extract coefficients and standard errors from models
results <- getModelResults(models)
}
}
\keyword{coefficient,}
\keyword{conjointTools,}
\keyword{error}
\keyword{standard}
