% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm.R
\name{mvrnorm}
\alias{mvrnorm}
\title{Simulate from a Multivariate Normal Distribution}
\usage{
mvrnorm(n = 1, mu, Sigma, tol = 1e-06, empirical = FALSE, EISPACK = FALSE)
}
\arguments{
\item{n}{the number of samples required.}

\item{mu}{a vector giving the means of the variables.}

\item{Sigma}{a positive-definite symmetric matrix specifying the
covariance matrix of the variables.}

\item{tol}{tolerance (relative to largest variance) for numerical lack
of positive-definiteness in \code{Sigma}.}

\item{empirical}{logical. If true, mu and Sigma specify the empirical
not population mean and covariance matrix.}

\item{EISPACK}{logical: values other than \code{FALSE} are an error.}
}
\value{
If \code{n = 1} a vector of the same length as \code{mu}, otherwise an
\code{n} by \code{length(mu)} matrix with one sample in each row.
}
\description{
Produces one or more samples from the specified multivariate normal
distribution. Copied from MASS::mvrnorm().
}
\examples{
Sigma <- matrix(c(10,3,3,2),2,2)
Sigma
var(mvrnorm(n = 1000, rep(0, 2), Sigma))
var(mvrnorm(n = 1000, rep(0, 2), Sigma, empirical = TRUE))

}
